+++ src/main/java/blackjack/domain/betting/Money.java
@@ -1,4 +1,4 @@
-package blackjack.domain.participant;
+package blackjack.domain.betting;
 
 import blackjack.utils.IntegerUtils;
